refactor: move new logic to gitlabci manager
