+++ lib/modules/manager/gitlabci-include/extract.ts
@@ -2,29 +2,20 @@ import is from '@sindresorhus/is';
 import { GlobalConfig } from '../../../config/global';
 import { logger } from '../../../logger';
 import { regEx } from '../../../util/regex';
-import { parseUrl } from '../../../util/url';
 import { parseSingleYaml } from '../../../util/yaml';
 import { GitlabTagsDatasource } from '../../datasource/gitlab-tags';
+import {
+  filterIncludeFromGitlabPipeline,
+  isGitlabIncludeProject,
+  isNonEmptyObject,
+} from '../gitlabci/common';
 import type {
   GitlabInclude,
-  GitlabIncludeComponent,
   GitlabIncludeProject,
   GitlabPipeline,
 } from '../gitlabci/types';
 import { replaceReferenceTags } from '../gitlabci/utils';
 import type { PackageDependency, PackageFileContent } from '../types';
-import {
-  filterIncludeFromGitlabPipeline,
-  isGitlabIncludeComponent,
-  isGitlabIncludeProject,
-  isNonEmptyObject,
-} from './common';
-
-// See https://docs.gitlab.com/ee/ci/components/index.html#use-a-component
-const componentReferenceRegex = regEx(
-  /(?<fqdn>[^/]+)\/(?<projectPath>.+)\/(.+)@(?<specificVersion>.+)/,
-);
-const componentReferenceLatestVersion = '~latest';
 
 function extractDepFromIncludeFile(
   includeObj: GitlabIncludeProject,
@@ -39,54 +30,6 @@ function extractDepFromIncludeFile(
     return dep;
   }
   dep.currentValue = includeObj.ref;
-  return dep;
-}
-
-function extractDepFromIncludeComponent(
-  includeComponent: GitlabIncludeComponent,
-  endpoint: string | undefined,
-): PackageDependency | null {
-  const componentReference = componentReferenceRegex.exec(
-    includeComponent.component,
-  )?.groups;
-  if (!componentReference) {
-    logger.debug(
-      { componentReference: includeComponent.component },
-      'Ignoring malformed component reference',
-    );
-    return null;
-  }
-  const projectPathParts = componentReference.projectPath.split('/');
-  if (projectPathParts.length < 2) {
-    logger.debug(
-      { componentReference: includeComponent.component },
-      'Ignoring component reference with incomplete project path',
-    );
-    return null;
-  }
-
-  const dep: PackageDependency = {
-    datasource: GitlabTagsDatasource.id,
-    depName: componentReference.projectPath,
-    depType: 'repository',
-    currentValue: componentReference.specificVersion,
-  };
-  if (dep.currentValue === componentReferenceLatestVersion) {
-    logger.debug(
-      { componentVersion: dep.currentValue },
-      'Ignoring component version',
-    );
-    dep.skipReason = 'unsupported-version';
-  }
-  const endpointUrl = parseUrl(endpoint);
-  if (endpointUrl && endpointUrl.hostname !== componentReference.fqdn) {
-    const registryUrl = `https://${componentReference.fqdn}`;
-    logger.debug(
-      { componentReference: includeComponent.component },
-      'Setting registry URL for external component reference',
-    );
-    dep.registryUrls = [registryUrl];
-  }
   return dep;
 }
 
@@ -120,29 +63,6 @@ function getAllIncludeProjects(data: GitlabPipeline): GitlabIncludeProject[] {
   return childrenData;
 }
 
-function getIncludeComponentsFromInclude(
-  includeValue: GitlabInclude[] | GitlabInclude,
-): GitlabIncludeComponent[] {
-  const includes = is.array(includeValue) ? includeValue : [includeValue];
-
-  // Filter out includes that dont have a file & project.
-  return includes.filter(isGitlabIncludeComponent);
-}
-function getAllIncludeComponents(
-  data: GitlabPipeline,
-): GitlabIncludeComponent[] {
-  const childrenData = Object.values(filterIncludeFromGitlabPipeline(data))
-    .filter(isNonEmptyObject)
-    .map(getAllIncludeComponents)
-    .flat();
-
-  // Process include key.
-  if (data.include) {
-    childrenData.push(...getIncludeComponentsFromInclude(data.include));
-  }
-  return childrenData;
-}
-
 export function extractPackageFile(
   content: string,
   packageFile?: string,
@@ -162,13 +82,6 @@ export function extractPackageFile(
         dep.registryUrls = [endpoint.replace(regEx(/\/api\/v4\/?/), '')];
       }
       deps.push(dep);
-    }
-    const includedComponents = getAllIncludeComponents(doc);
-    for (const includedComponent of includedComponents) {
-      const dep = extractDepFromIncludeComponent(includedComponent, endpoint);
-      if (dep) {
-        deps.push(dep);
-      }
     }
   } catch (err) /* istanbul ignore next */ {
     if (err.stack?.startsWith('YAMLException:')) {
