+++ lib/modules/manager/gitlabci/common.spec.ts
@@ -1,9 +1,34 @@
-import { isGitlabIncludeLocal } from './common';
+import { Fixtures } from '../../../../test/fixtures';
+import { parseSingleYaml } from '../../../util/yaml';
+import type { GitlabPipeline } from '../gitlabci/types';
+import { replaceReferenceTags } from '../gitlabci/utils';
+import {
+  filterIncludeFromGitlabPipeline,
+  isGitlabIncludeLocal,
+  isGitlabIncludeProject,
+  isNonEmptyObject,
+} from './common';
 
+const yamlFileMultiConfig = Fixtures.get('gitlab-ci.1.yaml');
+// TODO: use schema (#9610)
+const pipeline = parseSingleYaml<GitlabPipeline>(
+  replaceReferenceTags(yamlFileMultiConfig),
+);
 const includeLocal = { local: 'something' };
 const includeProject = { project: 'something' };
 
 describe('modules/manager/gitlabci/common', () => {
+  describe('filterIncludeFromGitlabPipeline()', () => {
+    it('returns GitlabPipeline without top level include key', () => {
+      expect(pipeline).toHaveProperty('include');
+      const filtered_pipeline = filterIncludeFromGitlabPipeline(pipeline);
+      expect(filtered_pipeline).not.toHaveProperty('include');
+      expect(filtered_pipeline).toEqual({
+        script: [null, null],
+      });
+    });
+  });
+
   describe('isGitlabIncludeLocal()', () => {
     it('returns true if GitlabInclude is GitlabIncludeLocal', () => {
       expect(isGitlabIncludeLocal(includeLocal)).toBe(true);
@@ -11,6 +36,26 @@ describe('modules/manager/gitlabci/common', () => {
 
     it('returns false if GitlabInclude is not GitlabIncludeLocal', () => {
       expect(isGitlabIncludeLocal(includeProject)).toBe(false);
+    });
+  });
+
+  describe('isGitlabIncludeProject()', () => {
+    it('returns true if GitlabInclude is GitlabIncludeProject', () => {
+      expect(isGitlabIncludeProject(includeProject)).toBe(true);
+    });
+
+    it('returns false if GitlabInclude is not GitlabIncludeProject', () => {
+      expect(isGitlabIncludeProject(includeLocal)).toBe(false);
+    });
+  });
+
+  describe('isNonEmptyObject()', () => {
+    it('returns true if not empty', () => {
+      expect(isNonEmptyObject({ attribute1: 1 })).toBe(true);
+    });
+
+    it('returns false if empty', () => {
+      expect(isNonEmptyObject({})).toBe(false);
     });
   });
 });
